fix: windows compile error

+++ core/include/niu2x/std_alias.h
@@ -197,7 +197,10 @@ inline TimeDuration time_diff(const TimePoint& t_new)
     return diff;
 }
 
-inline int64_t to_milliseconds(TimeDuration td) { return td * 1000; }
+inline int64_t to_milliseconds(TimeDuration td)
+{
+    return static_cast<int16_t>(td * 1000);
+}
 
 template <class T>
 inline const T nan = std::numeric_limits<T>::quiet_NaN();
